+++ Tensile/ClientWriter.py
@@ -1,5 +1,5 @@
 ################################################################################
-# Copyright 2016-2020 Advanced Micro Devices, Inc. All rights reserved.
+# Copyright 2016-2021 Advanced Micro Devices, Inc. All rights reserved.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -103,9 +103,6 @@ def main( config ):
   else:
     shutil_copy(
         os.path.join(globalParameters["SourcePath"], "FindHIP.cmake"),
-        globalParameters["WorkingPath"] )
-    shutil_copy(
-        os.path.join(globalParameters["SourcePath"], "FindHCC.cmake"),
         globalParameters["WorkingPath"] )
 
   ##############################################################################
@@ -631,9 +628,10 @@ def writeClientConfigIni(problemSizes, problemType, sourceDir, codeObjectFiles, 
                 param("code-object", os.path.join(sourceDir,coFile))
 
         param('results-file', resultsFileName)
-
-        if problemType.convolution and globalParameters["ConvolutionVsContraction"]:
+        convValidation = problemType.convolution and globalParameters["ConvolutionVsContraction"];
+        if convValidation:
             param('convolution-identifier', problemType.convolution.identifier())
+        param('performance-metric', globalParameters["PerformanceMetric"])
         param('problem-identifier', problemType.operationIdentifier)
         param('a-type',     problemType.aType.toEnum())
         param('b-type',     problemType.bType.toEnum())
@@ -648,6 +646,8 @@ def writeClientConfigIni(problemSizes, problemType, sourceDir, codeObjectFiles, 
         for problem in problemSizes.problems:
             for key,value in problemSizeParams(problemType, problem):
                 param(key,value)
+            if convValidation:
+              param('convolution-problem', problemType.convolution.identifier(problem))
 
         param("device-idx",               globalParameters["Device"])
 
@@ -681,8 +681,7 @@ def writeClientConfigIni(problemSizes, problemType, sourceDir, codeObjectFiles, 
         param("num-syncs-per-benchmark",  globalParameters["SyncsPerBenchmark"])
         param("use-gpu-timer",            globalParameters["KernelTime"])
         param("hardware-monitor",         globalParameters["HardwareMonitor"])
-        if globalParameters["ConvolutionVsContraction"]:
-            assert(problemType.convolution)
+        if convValidation:
             param("convolution-vs-contraction", globalParameters["ConvolutionVsContraction"])
         if not globalParameters["KernelTime"]:
             param("num-warmups", 1)