+++ Tensile/Source/client/include/ResultFileReporter.hpp
@@ -2,7 +2,7 @@
  *
  * MIT License
  *
- * Copyright 2019-2020 Advanced Micro Devices, Inc.
+ * Copyright 2019-2021 Advanced Micro Devices, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -46,7 +46,8 @@ namespace Tensile
 
             ResultFileReporter(std::string const& filename,
                                bool               exportExtraCols,
-                               bool               mergeSameProblems);
+                               bool               mergeSameProblems,
+                               PerformanceMetric  performanceMetric);
 
             virtual void reportValue_string(std::string const& key,
                                             std::string const& value) override;
@@ -66,11 +67,12 @@ namespace Tensile
             void reportValue(std::string const& key, T const& value);
             void mergeRow(std::unordered_map<std::string, std::string>& newRow);
 
-            CSVStackFile m_output;
-            std::string  m_solutionName;
-            bool         m_invalidSolution = false;
-            bool         m_extraCol;
-            bool         m_mergeSameProblems;
+            CSVStackFile      m_output;
+            std::string       m_solutionName;
+            bool              m_invalidSolution = false;
+            bool              m_extraCol;
+            bool              m_mergeSameProblems;
+            PerformanceMetric m_performanceMetric;
             // for extra columns
             std::string m_winnerSolution;
             int64_t     m_currSolutionIdx   = -1;