+++ Tensile/Source/client/source/ClientProblemFactory.cpp
@@ -2,7 +2,7 @@
  *
  * MIT License
  *
- * Copyright 2019-2020 Advanced Micro Devices, Inc.
+ * Copyright 2019-2021 Advanced Micro Devices, Inc.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -48,7 +48,8 @@ namespace Tensile
             , m_betaType(DataType::Float)
             , m_stridedBatched(args["strided-batched"].as<bool>())
             , m_highPrecisionAccumulate(args["high-precision-accumulate"].as<bool>())
-            , m_kernelLanguage(args["kernel-language"].as<Tensile::KernelLanguage>())
+            , m_kernelLanguage(args["kernel-language"].as<KernelLanguage>())
+            , m_performanceMetric(args["performance-metric"].as<PerformanceMetric>())
             , m_deterministicMode(args["deterministic-mode"].as<bool>())
             , m_arithmeticUnit(args["arithmetic-unit"].as<ArithmeticUnit>())
             , m_aStrides(args["a-strides"].as<std::vector<std::vector<size_t>>>())
@@ -96,6 +97,10 @@ namespace Tensile
                 m_betaType = args["beta-type"].as<DataType>();
 
             m_beta = DataInitialization::getValue<double>(args["init-beta"].as<InitMode>());
+
+            if(args["convolution-vs-contraction"].as<bool>())
+                m_convProblemSizes
+                    = args["convolution-problem"].as<std::vector<std::vector<size_t>>>();
 
             m_problems = createProblems();
         }
@@ -189,8 +194,12 @@ namespace Tensile
                 rv.back().setStridedBatched(m_stridedBatched);
                 rv.back().setHighPrecisionAccumulate(m_highPrecisionAccumulate);
                 rv.back().setKernelLanguage(m_kernelLanguage);
+                rv.back().setPerformanceMetric(m_performanceMetric);
                 rv.back().setDeterministicMode(m_deterministicMode);
                 rv.back().setArithmeticUnit(m_arithmeticUnit);
+
+                if(m_convProblemSizes.size())
+                    rv.back().setConvProblemSizes(m_convProblemSizes[i]);
             }
 
             return rv;