+++ Tensile/Tests/unit/test_conv_problem.py
@@ -1,3 +1,24 @@
+################################################################################
+# Copyright 2020-2021 Advanced Micro Devices, Inc. All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell cop-
+# ies of the Software, and to permit persons to whom the Software is furnished
+# to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IM-
+# PLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNE-
+# CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+################################################################################
+
 import logging,pytest
 from Tensile.SolutionStructs import Convolution,ConvProblem
 log =logging.getLogger("testlog")
@@ -12,7 +33,7 @@ def test_stride2x3():
     e= { 'n':64, 'c':256, 'h':20, 'w':14, 'k':1024, 'x':1, 'y':1, 'u':2, 'v':3 }
     ec = ConvProblem(e, conv)
     assert (ec.sizes == (5, 10, e['k'], e['n'], e['c']))
-    assert (ec.stridesA == (3, 28, -1, 71680))
+    assert (ec.stridesA == (3, 28, e['h'] * e['w'], 71680))
 
 def test_stride2x3_defaults():
     z={} # problemType definition
@@ -24,7 +45,7 @@ def test_stride2x3_defaults():
     e= { 'n':64, 'c':256, 'h':20, 'w':14, 'k':1024}
     ec = ConvProblem(e, conv)
     assert (ec.sizes == (5, 10, e['k'], e['n'], e['c']))
-    assert (ec.stridesA == (3, 28, -1, 71680))
+    assert (ec.stridesA == (3, 28, e['h'] * e['w'], 71680))
 
 @pytest.mark.skip(reason="no X filter, ZeroPadA has one entry and it is for Y filter")
 def test_pad_4x1():