+++ tuning/scripts/master_tuning_script.sh
@@ -1,5 +1,26 @@
 #!/usr/bin/env bash
 
+################################################################################
+# Copyright 2020-2021 Advanced Micro Devices, Inc. All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell cop-
+# ies of the Software, and to permit persons to whom the Software is furnished
+# to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IM-
+# PLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNE-
+# CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+################################################################################
+
 HELP=false
 TENSILE_CLIENT=new
 COUNT=false
@@ -19,7 +40,7 @@ Pre-Requisites:
   >=Anaconda 3.6 (or install python3.6 or higher, python3-pip/pip3, python3-yaml, python3-setuptools, python3-distutils,
       python3-venv, wheel, setuptools, pyyaml, msgpack, matplotlib, pandas, and numpy)
   >=llvm-6.0-dev, >=cmake3.5, zlib1g-dev
-  <=rocm3.0 stack for hcc, >=rocm3.3 stack for hip-clang
+  >=rocm3.3 stack for hip-clang
 
 About
 This is the master tuning script. It does the following: